+++ src/boutros/matsu/compiler.clj
@@ -89,8 +89,9 @@
   [m s]
   (str
     (apply str (apply sorted-set
-                      (for [[_ p] (re-seq #"(\b[a-zA-Z0-9]+):[a-zA-Z]" s) :when (not-any? #(= p %) #{"mailto" "sql" "bif"})]
-                        (str "PREFIX " p ": " (ns-or-error (keyword p) m) " " ))))
+                      (let [cleaned-string (clojure.string/replace s #"\".*\"" "")] ; as not to match inside quoted strings
+                      (for [[_ p] (re-seq #"(\b[a-zA-Z0-9]+):[a-zA-Z]" cleaned-string) :when (not-any? #(= p %) #{"mailto" "sql" "bif"})]
+                        (str "PREFIX " p ": " (ns-or-error (keyword p) m) " " )))))
     s))
 
 (defn- add-base